deploy-model-with-mms/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
multi-model-server==1.1.11
numpy==2.2.6
pillow==11.2.1
torch==2.7.0+cpu
torchvision==0.22.0+cpu